% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.notes.R
\name{add.notes}
\alias{add.notes}
\title{xxx}
\usage{
add.notes(series.list, xlim, ylim, names = NULL)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

