% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_predictions.R
\name{aggregate_predictions}
\alias{aggregate_predictions}
\title{aggregate predictions}
\usage{
aggregate_predictions(preds, totals = FALSE)
}
\arguments{
\item{preds}{predictions}

\item{totals}{logical}
}
\description{
aggregate predictions
}
