% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacen_v7.R
\docType{data}
\name{bacen_v7}
\alias{bacen_v7}
\title{BETS Time Series Metadata}
\format{A data frame with 39073 rows and 7 variables}
\source{
\url{http://www.bcb.gov.br}
}
\usage{
bacen_v7
}
\description{
A dataset containing many attributes of more 39,000 time series
}
\keyword{data}

