% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_final_ptv1.R
\docType{data}
\name{base_final_ptv1}
\alias{base_final_ptv1}
\title{BETS Time Series Metadata in portuguese}
\format{A data frame with 39073 rows and 7 variables}
\source{
\url{http://www.bcb.gov.br}
}
\description{
A dataset containing many attributes of more 39,000 time series
}
\keyword{data}

