% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbExpectAtop5.R
\name{bcbExpectATop5}
\alias{bcbExpectATop5}
\title{bcbExpectATop5}
\usage{
bcbExpectATop5(indicator = "IGP-DI", limit = 100,
  variables = c("tipoCalculo", "Media", "Mediana", "DesvioPadrao",
  "CoeficienteVariacao", "Minimo", "Maximo"), start, end)
}
\arguments{
\item{indicator}{A string. Available indicator.}

\item{limit}{A integer. A limint of data in request, top is 10000.}

\item{variables}{Possible options: "Media", "Mediana", "DesvioPadrao",
"CoeficienteVariacao", "Minimo", "Maximo".}

\item{start}{Initial date at which the data was projected, in ISO format.}

\item{end}{Final date at which the data was projected, in ISO format.}
}
\value{
A data.frame.
}
\description{
Annual Market Expectations Top5.
}
\note{
The available indicators are: IGP-DI, IGP-M, IPCA, Meta para taxa over-selic, Taxa de cambio.
}
\examples{
 # bcbExpectATop5()



}
