% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbExpectInf12.R
\name{bcbExpectInf12}
\alias{bcbExpectInf12}
\title{bcbExpectInf12}
\usage{
bcbExpectInf12(indicator = "IPC-FIPE", limit = 100, variables = "Media",
  start, end)
}
\arguments{
\item{indicator}{A string. Available indicator.}

\item{limit}{A integer. A limint of data in request, top is 10000.}

\item{variables}{Possible options: "Media", "Mediana", "DesvioPadrao",
"CoeficienteVariacao", "Minimo", "Maximo".}

\item{start}{Initial date at which the data was projected, in ISO format.}

\item{end}{Final date at which the data was projected, in ISO format.}
}
\value{
A data.frame.
}
\description{
Market expectations for inflation in the next 12 months
}
\note{
The available indicators are: IGP-DI, IGP-M, INPC, IPA-DI, IPA-M, IPCA, IPCA-15, IPC-FIPE.
}
\examples{
 # bcbExpectInf12()



}
