% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbExpectM.R
\name{bcbExpectM}
\alias{bcbExpectM}
\title{bcbExpectM}
\usage{
bcbExpectM(indicator = "IPCA-15", limit = 100, variables = c("Media",
  "Mediana", "DesvioPadrao", "CoeficienteVariacao", "Minimo", "Maximo",
  "numeroRespondentes", "baseCalculo"), start, end)
}
\arguments{
\item{indicator}{A string. Available indicator.}

\item{limit}{A integer. A limint of data in request, top is 10000.}

\item{variables}{Possible options: "Media", "Mediana", "DesvioPadrao",
"CoeficienteVariacao", "Minimo", "Maximo".}

\item{start}{Initial date at which the data was projected, in ISO format.}

\item{end}{Final date at which the data was projected, in ISO format.}
}
\value{
A data.frame.
}
\description{
Market Expectations with mensal reference.
}
\note{
The available indicators are: IGP-DI, IGP-M, INPC, IPA-DI, IPA-M, IPCA, IPCA-15, IPC-FIPE, Producao 
   industrial, Meta para taxa over-selic, Taxa de cambio .
}
\examples{
 # bcbExpectM()



}
