% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbExpectMtop5.R
\name{bcbExpectMTop5}
\alias{bcbExpectMTop5}
\title{bcbExpectMTop5}
\usage{
bcbExpectMTop5(indicator = "IGP-DI", limit = 100, variables = "Media",
  start, end)
}
\arguments{
\item{indicator}{A string. Available indicator.}

\item{limit}{A integer. A limint of data in request, top is 10000.}

\item{variables}{Possible options: "Media", "Mediana", "DesvioPadrao",
"CoeficienteVariacao", "Minimo", "Maximo".}

\item{start}{Initial date at which the data was projected, in ISO format.}

\item{end}{Final date at which the data was projected, in ISO format.}
}
\value{
A data.frame.
}
\description{
Monthly Market Expectations Top5.
}
\note{
The available indicators are: IGP-DI, IGP-M, IPCA, Meta para taxa over-selic, Taxa de cambio.
}
\examples{
 # bcbExpectMTop5()



}
