% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbExpectT.R
\name{bcbExpectT}
\alias{bcbExpectT}
\title{bcbExpectT}
\usage{
bcbExpectT(indicator = "PIB Total", limit = 100, variables = "Media",
  start, end)
}
\arguments{
\item{indicator}{A string. Available indicator.}

\item{limit}{A integer. A limint of data in request, top is 10000.}

\item{variables}{Possible options: "Media", "Mediana", "DesvioPadrao",
"CoeficienteVariacao", "Minimo", "Maximo".}

\item{start}{Initial date at which the data was projected, in ISO format.}

\item{end}{Final date at which the data was projected, in ISO format.}
}
\value{
A data.frame.
}
\description{
Quarterly Market Expectations.
}
\note{
The available indicators are: PIB Agropecuario, PIB Industrial, PIB Serviços e PIB Total.
}
\examples{
 # bcbExpectT()



}
