% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.addin_pt.R
\name{BETS.addin_pt}
\alias{BETS.addin_pt}
\title{BETS search}
\usage{
BETS.addin_pt()
}
\description{
An interface for searching time series with possibility 
to extract the data in different extensions.
}
