% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.chart.R
\name{BETS.chart}
\alias{BETS.chart}
\title{Create a chart with BETS aesthetics}
\usage{
BETS.chart(ts, file = NULL, open = TRUE, lang = "en", params = NULL)
}
\arguments{
\item{ts}{A \code{character} or a \code{ts} object. A custom time series or the name of a pre-defined series. A complete list of names is under the 'Details' section.}

\item{file}{A \code{character}. The whole path, including a custom name, for the output (an image file). The default value is 'graphs//parameter_ts' (the 'graphs' directory is under the BETS installation directory).}

\item{open}{A \code{boolean}. Whether to open the file containing the chart.}

\item{lang}{A \code{character}. The language. For now, only 'en' (english) is available.}

\item{params}{A \code{list}. Parameters for drawing custom charts. See the 'details' section.}
}
\value{
If parameter \code{file} is not set by the user, the chart will be shown at the standard R ploting area. Otherwise, it is going to be saved on your computer.
}
\description{
Create a chart with a professional look, using a pre-defined BETS series or a custom series.
}
\details{
\bold{Names of pre-defined charts:}

\tabular{lll}{
 VALUE \tab DESCRIPTION \tab CODE \cr
 \bold{'animal_spirits'} \tab \tab (*) \cr
 \bold{'iie_br'} \tab \tab ST_100.0 \cr
 \bold{'ei_vars'} \tab \tab (*) \cr
 \bold{'ei_comps'} \tab  \tab (*) \cr
 \bold{'gdp_vars'} \tab \tab (*) \cr
 \bold{'misery_index} \tab \tab 13522 plus 24369 \cr
 \bold{'gdp_comps'} \tab  \tab (*) \cr
 \bold{'gdp_unemp'} \tab  \tab 22109 and 24369 \cr
 \bold{'conf_lvl'} \tab  \tab (*) \cr
 \bold{'inst_cap'} \tab  \tab (*) \cr
 \bold{'lab_mrkt'} \tab  \tab (*) \cr
 \bold{'transf_ind'} \tab  \tab (*) \cr
 \bold{'servc'} \tab  \tab (*) \cr
 \bold{'constr'} \tab  \tab (*) \cr
 \bold{'retail'} \tab  \tab (*) \cr
 \bold{'consm'} \tab \tab (*) 
}

(*) Not in BETS databases yet. But you can find it in .csv files saved under your BETS installation directory.

\bold{Parameters for custom charts:}
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

