% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.corrgram.R
\name{BETS.corrgram}
\alias{BETS.corrgram}
\title{xxx}
\usage{
BETS.corrgram(ts, lag.max = 12, type = "correlation", mode = "simple",
  ci = 0.95, style = "plotly", knit = F)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

