% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.deflate.R
\name{BETS.deflate}
\alias{BETS.deflate}
\title{xxx}
\usage{
BETS.deflate(ts, deflator, type = "index")
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

