% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.dummy.R
\name{BETS.dummy}
\alias{BETS.dummy}
\title{Create dummy to monthly ts object}
\usage{
BETS.dummy(start = NULL, end = NULL, frequency = 12, year = NULL,
  month = NULL, quarter = NULL, date = NULL, from = NULL, to = NULL)
}
\arguments{
\item{start}{the time of the first observation. Either a vector of 
two integers, which specify a natural time unit and a (1-based) 
number of samples into the time unit.}

\item{end}{the time of the last observation, specified in the same way as start}

\item{frequency}{the number of observations per unit of time (12 in defaut).}

\item{year}{a \code{integer},or a \code{seq} or a \code{vector} that 
contains the years. Which year or years will be created the dummys}

\item{month}{a \code{integer} or a \code{vector} of size 12 that 
contains the months. For which month or months will be created the dummys}

\item{quarter}{a \code{integer} or a \code{vector} of size 4 that contains
the quarters. For which quarter or quarters will be created the dummys}

\item{date}{a \code{list} with the date objects.}

\item{from}{the starting  value of the sequence. Of 
length 1 unless just from is supplied as an unnamed argument.}

\item{to}{the maximal value of the sequence.}
}
\value{
A monthly or a quarterly \code{ts}.
}
\description{
Given a informations about the time and dates, return a monthly 
or quarterly ts with dummies.
}
\details{
the BETS.dummy works for the monthly or a quarterly \code{ts}.
}
\examples{

#1 from a specific date to another specific date
BETS.dummy(start = c(2000,1),end = c(2012,5),frequency = 12,from = c(2005,1),to = c(2006,12))


#Other options that may be helpful:

#over a month equal to 1
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = c(5,12))

#Months equal to 1 only for some year
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = 5, year = 2010)
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = 8, year = 2002)

#Months equal to 1 only for some years
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = 5, year = 2005:2007)
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = 3, year = c(2005,2007))
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, month = 5:6, year = c(2005,2007))

#specific dates
BETS.dummy(start = c(2000,1), end = c(2012,5), frequency = 12, date = list(c(2010,1)))
BETS.dummy(start = c(2000,1), end = c(2012,5), 
    freq = 12, date = list(c(2010,9), c(2011,1), c(2000,1)) )


}
\seealso{
\code{\link[stats]{ts}}, \code{\link[BETS]{BETS.dummy}}
}

