% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.grnn.R
\name{BETS.grnn}
\alias{BETS.grnn}
\title{xxx}
\usage{
BETS.grnn(train, test, sigma, step = 0.1, select = TRUE, names = NA)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

