% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.grnn.test.R
\name{BETS.grnn.test}
\alias{BETS.grnn.test}
\title{xxx}
\usage{
BETS.grnn.test(results, test.set, select = TRUE)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

