% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.grnn.train.R
\name{BETS.grnn.train}
\alias{BETS.grnn.train}
\title{xxx}
\usage{
BETS.grnn.train(train, sigma, step = 0.1, select = TRUE, names = NA)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

