% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.normalize.R
\name{BETS.normalize}
\alias{BETS.normalize}
\title{xxx}
\usage{
BETS.normalize(series, mode)
}
\arguments{
\item{mode}{xxx}

\item{ts}{xxx}

\item{lag.max}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

