% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.plot.R
\name{BETS.plot}
\alias{BETS.plot}
\title{xxx}
\usage{
BETS.plot(ts, start = NULL, end = NULL, ylab = NULL, type = "series")
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

