% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.predict.R
\name{BETS.predict}
\alias{BETS.predict}
\title{xxx}
\usage{
BETS.predict(..., actual = NULL, main = "", ylab = "", xlim = NULL,
  style = "dygraphs", knit = F)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

