% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.R
\docType{package}
\name{BETS}
\alias{BETS}
\alias{BETS-package}
\title{BETS: A package for obtaining and analysing thousands of Brazilian economic time series.}
\description{
The Brazilian Economic Time Series (BETS) package provides access and
information about the most important Brazilian economic time series.

These series are created by three influential centers: the Central Bank of
Brazil (BCB), the Brazilian Institute of Geography and Statistics (IBGE)
and the Brazilian Institute of Economics, from the Getulio Vargas Foundation
(FVG-IBRE). Currently, there are more than 18.640 available time series, most
of them free of charge. Besides providing access to this vast database, the
package allows the user to interact with data in an easy and friendly way.

For instance, the user can search for a time series using keywords. More
importantly, it installs several consecrated packages for time series
analysis, giving the user the option to perform a complete analysis without
having to worry about installing and loading other packages. In a near
future, the authors will publish a series of R exercises to be solved with
BETS and its statiscal/econometrical tools, therefore helping the user to
understand the behavior of brazilian time series.
}
\note{
The authors would like to thank the support by the Getulio
Vargas Foundation (FGV) and make it clear that all
data in the package is in public domain. The rights of all centers from
which the series are taken are maintained. We reaffirm that BETS
is mainly intended for academic usage.
}
\author{
Pedro Costa Ferreira \email{pedro.guilherme@fgv.br},
        Jonatha Costa \email{jonatha.costa@fgv.br},
        Talitha Speranza \email{talitha.speranza@fgv.br}
}
