% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.report.R
\name{BETS.report}
\alias{BETS.report}
\title{xxx}
\usage{
BETS.report(code = 21864, model = "SARIMA", parameters = NULL)
}
\arguments{
\item{model}{xxx}

\item{params}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

