% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.std_resid.R
\name{BETS.std_resid}
\alias{BETS.std_resid}
\title{xxx}
\usage{
BETS.std_resid(model, alpha = 0.95)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

