% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.t_test.R
\name{BETS.t_test}
\alias{BETS.t_test}
\title{Significance of parameter an Arima model}
\usage{
BETS.t_test(arima, n = 0, alpha = 0.05)
}
\arguments{
\item{arima_model}{Arima model used}

\item{n_x}{Number of variables Exogenous}
}
\value{
Objeto do tipo \code{list}
}
\description{
Performs the test of significance of the parameter an 
Arima model
}
\examples{
data("AirPassengers")
fit.air<- Arima(AirPassengers,order = c(1,1,1), 
    seasonal = c(1,1,1), method ="ML",lambda=0)
    summary(fit.air)

significance test for model SARIMA(1,1,1)(1,1,1)_12
t.test(arima_model = fit.air)


}
\author{
Daiane Mattos \email{daiane.mattos@fgv.br}
}

