% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.ur_test.R
\name{BETS.ur_test}
\alias{BETS.ur_test}
\title{xxx}
\usage{
BETS.ur_test(..., mode = "DF", level = "5pct")
}
\arguments{
\item{mode}{xxx}

\item{ts}{xxx}

\item{lag.max}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

