% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSchart.R
\name{BETSchart}
\alias{BETSchart}
\title{xxx}
\usage{
BETSchart(alias)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

