% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSdummy.R
\name{BETSdummy}
\alias{BETSdummy}
\title{Create dummy to monthly and/or \code{ts} object}
\usage{
BETSdummy(star,end,freq = 12,year,month,quarter,date,from,to)
}
\arguments{
\item{start}{the time of the first observation. Either a single number or
a vector of two integers, which specify a natural time unit and a (1-based) 
number of samples into the time unit.}

\item{end}{the time of the last observation, specified in the same way as start}

\item{freq}{the number of observations per unit of time (12 in defaut).}

\item{year}{A \code{integer},or a \code{seq} or a \code{vector} that 
contains the years.}

\item{month}{A \code{integer} or a \code{vector} of size 12 that 
contains the months.}

\item{quarter}{A \code{integer} or a \code{vector} of size 4 that contains
the quarters.}

\item{date}{A \code{list} with the date objects.}

\item{from, }{to the starting and (maximal) end values of the sequence. Of 
length 1 unless just from is supplied as an unnamed argument.}
}
\value{
A monthly \code{ts}.
}
\description{
Given a informations about the time and dates, return a monthly 
and/or quarterly \code{ts} with dammies.
\code{ts}
}
\details{
the BETSdammy works for the monthly or a quarterly \code{ts}.
}
\examples{

#1 from a specific date to another specific date
dummy(start = c(2000,1),end = c(2012,5),freq = 12,from = c(2005,1),to = c(2006,12))


#Other options that may be helpful:

#over a month equal to 1
dummy(start = c(2000,1), end = c(2012,5), freq = 12, month = c(5,12))

# Months equal to 1 only for some years
dummy(start = c(2000,1), end = c(2012,5), freq = 12, month = 5, year = 2005)
dummy(start = c(2000,1), end = c(2012,5), freq = 12, month = 5, year = 2005:2007)
dummy(start = c(2000,1), end = c(2012,5), freq = 12, month = 5, year = c(2005,2007))
dummy(start = c(2000,1), end = c(2012,5), freq = 12, month = 5:6, year = c(2005,2007))

#specific dates
dummy(start = c(2000,1), end = c(2012,5), freq = 12, date = list(c(2010,1)))
dummy(start = c(2000,1), end = c(2012,5), 
    freq = 12, date = list(c(2010,9), c(2011,1), c(2000,1)) )

}

