% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSget.R
\name{BETSget}
\alias{BETSget}
\title{Get a complete time series from a BETS database}
\usage{
BETSget(code, data.frame = FALSE)
}
\arguments{
\item{code}{A \code{character}. The unique code that references the time series. This code can be obtained by using the \code{\link{BETSsearch}} function.}
}
\value{
A \code{\link[stats]{ts}} (time series) object containing the desired series.
}
\description{
Extracts a complete time series from either the Central Bank of Brazil (BCB), the Brazilian Institute of Geography and Statistics (IBGE) or the Brazilian Institute of Economics (FGV/IBRE).
}
\note{
Due to the significant size of the databases, it could take a while to retrieve the values. However, it shouldn't take more than 90 seconds.
}
\examples{

 # Anual series: GDP at constant prices, in R$ (brazilian reais)
 BETSget(1208)
 
 # International reserves - Cash concept 
 x <- BETSget("3543")
 plot(x)
 
 # Exchange rate - Free - United States dollar (purchase)
 k <- BETSget(3691)
 requires(seasonal)
 m <- seas(k)
 plot(m)
 
}
\seealso{
\code{\link[stats]{ts}}, \code{\link[BETS]{BETSsearch}} and \code{\link[seasonal]{seas}}
}
\keyword{get}

