% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSreport.R
\name{BETSreport}
\alias{BETSreport}
\title{xxx}
\usage{
BETSreport(model = "SARIMA", parameters = NULL)
}
\arguments{
\item{model}{xxx}

\item{params}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

