% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsave.sas.R
\name{BETSsave.sas}
\alias{BETSsave.sas}
\title{Export a time series to SAS}
\usage{
BETSsave.sas(code, data = NULL, file.name = "series")
}
\arguments{
\item{code}{An \code{integer}. The unique identifier of the series within the BETS database.}

\item{data}{A \code{data.frame} or a \code{ts}. Contains the data to be written. If \code{data} is supplied, the BETS database will not be searched.}

\item{file.name}{A \code{character}. The name of the output file. The default is 'series.sas'.}
}
\value{
None
}
\description{
Writes a time series to a .sas (SAS) file.
}

