% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsave.spss.R
\name{BETSsave.spss}
\alias{BETSsave.spss}
\title{Exportar series em extensao .sps (SPSS)}
\usage{
BETSsave.spss(x, codigo, file.name = "teste")
}
\arguments{
\item{x}{obejto}

\item{codigo}{que referencia a serie temporal}

\item{file.name}{nome com o qual o arquivo sera exportado}
}
\value{

}
\description{

}
\note{

}
\examples{

}
\references{

}

