% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsearch.R
\name{BETSsearch}
\alias{BETSsearch}
\title{Search for a Brazilian Economic Time Series}
\usage{
BETSsearch(name, src, periodicity, unit, code, view = TRUE, database)
}
\arguments{
\item{name}{A \code{character}. The complete name or a part of the name of the series.}

\item{src}{A \code{character}. The source of the series. See the 'Details' section for a list of the available sources.}

\item{periodicity}{A \code{character}. The periodicity of the series. See the 'Details' section for a list of possible values.}

\item{unit}{A \code{character}. The unit of measurement of the data. See the 'Details' section for a list of possible values.}

\item{code}{An \code{integer}. The index of the series within the database.}

\item{view}{A \code{boolean}. The default is \code{TRUE}. If set to \code{FALSE}, the results are NOT going to be shown.}

\item{database}{A \code{character}. The database from which the series must be retrieved. The default is \code{bacen}. See the 'Details' section for a list of available databases.}
}
\value{
A \code{list} that can be interpreted as a \code{data.frame}. The fields are described below.

\tabular{ll}{
 code \tab The code/index of the series within the database \cr
 description \tab The description of the series \cr
 periodicity \tab The periodicity of the series \cr
 start \tab Starting date of the series \cr
 source \tab The source of the series \cr
 unit \tab The unit of measurement of the data
}
}
\description{
Searches the BETS databases for a time series by its name, source, periodicity, code, data, unit of measurement and database name.
}
\details{
\itemize{

\item{ Possible values for the parameter \code{src}:
   \tabular{ll}{
     IBGE \tab Brazilian Institute of Geography and Statistics \cr
     BCB \tab Central Bank of Brazil \cr
     BCB e Deban \tab desc \cr
     BCB e FGV \tab Central Bank of Brazil and Getúlio Vargas Foundation \cr
     BCB-Depin \tab desc \cr
     BCB-Derin \tab desc \cr
     BCB-Desig \tab desc \cr
     BCB-Secre \tab desc \cr
     BCB-Demab \tab desc \cr
     BCB-Denor \tab desc \cr
     FGV \tab Getúlio Vargas Foundation \cr
     Sisbacen e Abecip \tab desc \cr
     BCB-Depec \tab desc
   }
}

\item{ Possible values for the parameter \code{periodicity}:
   \tabular{ll}{
     A - \tab anual data \cr
     M - \tab monthly data \cr
     Q - \tab quaterly data \cr
     W - \tab weekly data \cr
     D - \tab daily data 
   }
}

\item{ Possible values for the parameter \code{unit}:
   \tabular{ll}{
     R$ - \tab brazilian reais \cr
     $ - \tab US dolars \cr
    \% - \tab percentage 
   }
}

\item{ Possible values for the parameter \code{database}:
   \tabular{ll}{
     bacen - \tab Central Bank of Brazil
   }
}

}
}
\note{
This function uses \code{\link[sqldf]{sqldf}} for optimization.
}
\examples{

BETSsearch(name="sales")
# Output: BETS-package: 55 of 12981 time series !

BETSsearch(code= 4500)
# Output: BETS-package: DONE!

BETSsearch(src="Denor")
# Output: BETS-package: 1 of 12981 time series !

BETSsearch(periodicity="A")
# Output: BETS-package: 2308 of 12981 time series!

}
\references{
Central Bank of Brazil. \href{https://www3.bcb.gov.br/sgspub/localizarseries/localizarSeries.do?method=prepararTelaLocalizarSerie}{Time Series Management System - v2.1}
}
\keyword{search}

