% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsources.R
\name{BETSsources}
\alias{BETSsources}
\title{Display a list of sources available at BETS package}
\usage{
BETSsources()
}
\description{
Display a list of sources available at BETS package in console. The numbers of
sources will increase wiht new versions of the package.
}
