% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.add_basic.R
\name{chart.add_basic}
\alias{chart.add_basic}
\title{Create a chart of the Unitary Labor Cost time series}
\usage{
chart.add_basic(ts, xlim = NULL, ylim = NULL, type = "lines",
  title = "", subtitle = "", col = "firebrick4", arr.size = NULL,
  arr.pos = "v", leg.pos = "top", trend = FALSE)
}
\arguments{
\item{ts}{A \code{ts}. the ts object.}

\item{xlim}{A \code{numeric vector}. x axis limits.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{type}{A \code{character}. The type of of plot (lines).}

\item{title}{A \code{character}. The plot title.}

\item{subtitle}{A \code{character}. The plot subtitle.}

\item{col}{A \code{character}. Color.}

\item{arr.size}{A \code{vector}.}

\item{arr.pos}{A \code{vector}.}

\item{leg.pos}{A \code{vector}.}

\item{trend}{A \code{boolean}.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 11777
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
