% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.add_extra.R
\name{chart.add_extra}
\alias{chart.add_extra}
\title{Create a chart of the Unitary Labor Cost time series}
\usage{
chart.add_extra(ts, ylim = NULL, xlim = NULL, col = "firebrick3",
  arr.size = NULL, arr.pos = "v", leg.pos = "top", leg.text = "",
  main.type = "lines")
}
\arguments{
\item{ts}{A \code{ts}. the ts object.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{xlim}{A \code{numeric vector}. x axis limits.}

\item{col}{A \code{character}. Color.}

\item{arr.size}{A \code{}.}

\item{arr.pos}{A \code{}.}

\item{leg.pos}{A \code{}.}

\item{leg.text}{A \code{}.}

\item{main.type}{A \code{}.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 11777
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
