% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.series.R
\name{check.series}
\alias{check.series}
\title{xxx}
\usage{
check.series(ts, message = NULL)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

