\name{Code.of.series}
\alias{Code.of.series}
\title{Localização do código que referência as séries temporais dentro da base de dados do pacote FGVTS}
\usage{
code.of.series(n)
}
\arguments{
\item{n}{ Número inteiro, dimensão}
}
\description{Função para pesquisa dos códigos das séries temporais contidas no pacote FGVTS
}
\author{FGV - fundação Getúlio Vargas}
\examples{
code.of.series(20)
}
\keyword{ts}
