% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrgram.R
\name{corrgram}
\alias{corrgram}
\title{xxx}
\usage{
corrgram(ts, lag.max = 10, type = "correlation", mode = "simple",
  ci = 0.95)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

