% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.animal_spirits.R, R/draw.gdp_unemp.R
\name{draw.animal_spirits}
\alias{draw.animal_spirits}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.animal_spirits()

draw.animal_spirits()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.

An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189

Creates a plot of series 4189
}

