% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ei_comps.R
\name{draw.ei_comps}
\alias{draw.ei_comps}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.ei_comps()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}
