% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.generic.R
\name{draw.generic}
\alias{draw.generic}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.generic(ts, style, params)
}
\arguments{
\item{ts}{aaaa}

\item{style}{aaa}

\item{params}{aaa}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
