% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.lab_mrkt.R
\name{draw.lab_mrkt}
\alias{draw.lab_mrkt}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.lab_mrkt()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}
