% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.misery_rate.R
\name{draw.misery_rate}
\alias{draw.misery_rate}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.misery_rate()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}

