% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ulc.R
\name{draw.ulc}
\alias{draw.ulc}
\title{xxx}
\usage{
draw.ulc(file, start = NULL, ylim = NULL)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}

