% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bets.R
\name{find_bets}
\alias{find_bets}
\title{helper fun to find bets}
\usage{
find_bets(arviot, totals = TRUE)
}
\arguments{
\item{arviot}{data to ...}

\item{totals}{logical}
}
\description{
helper fun to find bets
}
