% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{helper fun to fit models}
\usage{
fit_models(pin_data, configs, min_games = 60)
}
\arguments{
\item{pin_data}{data to fit}

\item{configs}{lasso configs}

\item{min_games}{kuinka paljon minimissaan peleja}
}
\description{
helper fun to fit models
}
