% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multimixture_model.R
\name{fit_multimixture_model}
\alias{fit_multimixture_model}
\title{fit multimixture model}
\usage{
fit_multimixture_model(games)
}
\arguments{
\item{games}{games dataset}
}
\description{
fit multimixture model
}
