% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.series.bacen.R
\name{get.series.bacen}
\alias{get.series.bacen}
\title{A function to extract BACEN series using their API}
\usage{
get.series.bacen(x, from = "", to = "", save = "")
}
\arguments{
\item{x}{Bacen series numbers. Either an integer or a numeric vector.}

\item{from}{A string specifying where the series shall start.}

\item{to}{A string specifying where the series shall end.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract BACEN series using their API
}
\author{
Fernando Teixeira \email{fernando.teixeira@fgv.br} and Jonatha Azevedo 
\email{jonatha.costa@fgv.br}
}
\keyword{bacen}
