% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buchdal_data.R
\name{get_extra_leagues}
\alias{get_extra_leagues}
\title{get extra leagues data from buchdal}
\usage{
get_extra_leagues(URL, liigat)
}
\arguments{
\item{URL}{url code}
}
\description{
get extra leagues data from buchdal
}
