% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fbref_data.R
\name{get_fbref_data}
\alias{get_fbref_data}
\title{get fbref data}
\usage{
get_fbref_data(countries, tiers, seasons)
}
\arguments{
\item{countries}{a vector of countries}

\item{tiers}{a vector of tiers}

\item{seasons}{a vector of years}
}
\value{
a tibble containing data.
}
\description{
get fbref data
}
