% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_buchdata.R
\name{get_historical_buchdata}
\alias{get_historical_buchdata}
\title{get historical data of buchal for all years}
\usage{
get_historical_buchdata(urls)
}
\arguments{
\item{urls}{for mainleagues}
}
\description{
get historical data of buchal for all years
}
