% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pinnacle_odds.R
\name{get_pinnacle_odds}
\alias{get_pinnacle_odds}
\title{get odds from pinnacle}
\usage{
get_pinnacle_odds(liigat)
}
\arguments{
\item{liigat}{vector of league codes}
}
\description{
get odds from pinnacle
}
