% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pinnacle_odds.R
\name{HaePinnaclenKertoimet}
\alias{HaePinnaclenKertoimet}
\title{get odds from pinnacle}
\usage{
HaePinnaclenKertoimet(liigat)
}
\arguments{
\item{liigat}{vector of league codes}
}
\description{
get odds from pinnacle
}
