% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIE-Br-Expectativa.R
\name{IIE-Br- Expectativa}
\alias{Expectativa}
\alias{IIE-Br-}
\alias{IIE-Br- Expectativa}
\title{IIE-Br- Expectativa}
\description{
IIE-Br- Expectativa
}
\details{
O \code{IIE-Br- Expectativa} e um indicador que objetiva calcular o grau de 
dispersão nas expectativas de mercado, calculadas pelo Banco Central 
e pesquisas de sondagem, calculados Empresariais, calculadas pela FGV/IBRE".

O \code{IIE-Br- Expectativa} e indicador composto por dois outros 
indicadores.O Prmeiro e o indicador composto por informacoes do Banco 
central que contruido atraves da media das expectativas de mercado do IPCA,
Taxa Selic, e a Formacao primeria do Fiscal. O Segundo e o indicador que 
contemplas as inforrmacaos das sondagens empresarias calculados pela 
FGV|IBRE, esse indicador e construido atraves da ponderacao da Sonondagen 
da Industria, Sondagem do Comercio, Sondagem de Servicos e Sondagem da 
Construcao com suas respectivas areas nas contas nacionais.
}
\author{
Pedro G. C. Ferreira  \email{pedro.guilherme@fgv.br},
        Anna Carolina S. Barros \email{anna.barros@fgv.br},
        Bruno R de Miranda Neto \email{bruno.neto@fgv.br},
        Itaiguara de Oliveira Bezerra  \email{itaiguara.bezerra@fgv.br}.
}
\references{
\url{https://pedroferreira.shinyapps.io/incerteza/}
}
\keyword{Expectativa}
\keyword{IIE-Br-}

