% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIE_Br_media.R
\name{IIE_Br_media}
\alias{IIE_Br_media}
\title{Uncertainty Indicator for the Brazilian Economy - Media}
\format{A \code{\link[stats]{ts}} object with 185 observations.}
\source{
\href{http://portalibre.fgv.br/}{Brazilian Institute of Economics (FGV/IBRE)}
}
\description{
O \code{IIE-Br-Midia} e uma medida de incerteza baseada nas publicacoes
dos principais jornais do pais. Para ser calculado, ele leva em conta a 
frequencia de noticias contendo determinados termos, que remetem a incerteza
economica. Esse indice foi inspirado em Baker et al.(2015) e Alexopoulos and
Cohen (2009), que construiram indices semelhantes para os Estados Unidos. 
Os autores mostram que periodos de maior incerteza tendem a ser representados 
em noticias contendo os fatos causadores da incerteza. Essas noticias por 
consequencia, disseminam a sensação de incerteza para consumidores e 
produtores, e influenciam suas tomadas de decisoes. Filho(2014) tambem cria 
um indice semelhante para a economia brasileira.

Para ser construido, o \code{IIE-Br-Midia}  utilizou diferentes bases de 
dados, de forma a contemplar os principais meios de disseminação de noticias.
A primeira base, utilizou dados das redes sociais Twitter e Facebook. A 
segunda diz respeito, aos sites dos jornais, isto é, das versoes online 
dos jornais e por fim, a ultima base de dados foram as versoes impressas dos
jornais, em formato digital. Todo o conjunto de dados foi capturado e 
manipulado, computacionalmente, usando a linguagem R.
}
\author{
Pedro Costa Ferreira  \email{pedro.guilherme@fgv.br},
        Anna Carolina S. Barros \email{anna.barros@fgv.br},
        Bruno R de Miranda Neto \email{bruno.neto@fgv.br},
        Itaiguara de Oliveira Bezerra  \email{itaiguara.bezerra@fgv.br}.
}
\references{
A shiny app with IIE-Br-expectations plots can be found 
             \href{here}{https://pedroferreira.shinyapps.io/incerteza/}
}
\keyword{media}
\keyword{uncertainty,}

