\name{informations}
\alias{informations}
\title{Location code , name and size of the database}
\usage{
search.fgv2(code,name,n)
}
\arguments{
\item{code}{code reference series}
}
\item{name}{Name of series}

\description{ function for search names of times series
}
\author{FGV -  Getulio Vargas Foundation Time Series}
\examples{
name.of.series(20)
}
\keyword{ts}
