% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_buch_fbref.R
\name{join_buch_fbref}
\alias{join_buch_fbref}
\title{Join Buch Data and Fbref Data}
\usage{
join_buch_fbref(buch_data, fbref_data)
}
\arguments{
\item{buch_data}{data}

\item{fbref_data}{data}
}
\value{
A tibble containing the joined data.
}
\description{
This function joins data using a data.table join.
}
