% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelly_bet.R
\name{kelly_bet}
\alias{kelly_bet}
\title{calculate kelly bet}
\usage{
kelly_bet(ev, kerroin, max_limit)
}
\arguments{
\item{ev}{expected value}

\item{kerroin}{pelin kerroin}

\item{max_limit}{max bet}
}
\description{
calculate kelly bet
}
