% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Format and show a console message.}
\usage{
msg(..., skip_before = TRUE, skip_after = FALSE)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[base]{message}}}

\item{skip_before}{A \code{boolean}. Indicates if a line should be skipped before the message.}

\item{skip_after}{A \code{boolean}. Indicates if a line should be skipped after the message.}
}
\value{
None
}
\description{
Customizes a message and shows it in the console.
}

