\name{Name.of.series}
\alias{Name.of.series}
\title{Nome das séries temporais dentro do pacote FGVTS}
\usage{
name.of.series(n)
}
\arguments{
\item{n}{ Número inteiro}
}
\description{Função para pesquisa dos nomes das séries temporais contidas no pacote FGVTS
}
\author{FGV - fundação Getúlio Vargas}
\examples{
name.of.series(20)
}
\keyword{ts}
