% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a time series}
\usage{
normalize(series, mode = "scale")
}
\arguments{
\item{series}{A \code{ts} object or a \code{ts list}. The series to be normalized.}

\item{mode}{A \code{character}. The normalization method. Set this parameter to 'maxmin' to map series values to values between 0 and 1. Alternatively, set this parameter to 'scale' to standardize (substract the mean and divide by the standard deviation).}
}
\value{
A \code{ts} object or a \code{ts list}. The normalized series.
}
\description{
Normalizes a time series, either by stardization or by mapping to values between 0 and 1.
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
