% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_with_fits.R
\name{pred_with_fits}
\alias{pred_with_fits}
\title{predict with  fits}
\usage{
pred_with_fits(fits, totals = TRUE, pin_odds)
}
\arguments{
\item{fits}{goalmodel object}

\item{totals}{logical}

\item{pin_odds}{dataset}
}
\description{
predict with  fits
}
