% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_1x2.R
\name{predict_1x2}
\alias{predict_1x2}
\title{fit multimixture model}
\usage{
predict_1x2(model, data)
}
\arguments{
\item{model}{goalmodel object}

\item{data}{dataset}
}
\description{
fit multimixture model
}
