% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_buch_with_pinnacle.R
\name{replace_buch_with_pinnacle}
\alias{replace_buch_with_pinnacle}
\title{replaces buchdal names with pinnacle}
\usage{
replace_buch_with_pinnacle(data, team1, team2)
}
\arguments{
\item{data}{account name}

\item{team1}{account name}

\item{team2}{account name}
}
\description{
replaces buchdal names with pinnacle
}
