% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bets.R
\name{save_bets}
\alias{save_bets}
\title{helper fun to save bets}
\usage{
save_bets(to_save, arviot = TRUE)
}
\arguments{
\item{to_save}{data to ...}

\item{arviot}{vai bets kumpi tallennetaan}
}
\description{
helper fun to save bets
}
