% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.sidra.R
\name{search.sidra}
\alias{search.sidra}
\title{Search for Sidra Series}
\usage{
search.sidra(x, browse = FALSE)
}
\arguments{
\item{x}{Either a character or a numeric. If character, function searches the Sidra metadata. If a numeric argument is provided the descriptions of the given table are seached .}

\item{browse}{Logical. If browse is set to TRUE, the description table opens in your browser for better visualization.}
}
\description{
Searches the Sidra databases for a series by its description or a given table descriptions.
}
\examples{
\dontrun{
search.sidra("pib")
sear.sidra(1248)
}
}
\keyword{IBGE}
\keyword{sidra}
