% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_notification.R
\name{send_notification}
\alias{send_notification}
\title{helper fun to send bets}
\usage{
send_notification(betit, totals = FALSE)
}
\arguments{
\item{betit}{betit}
}
\description{
helper fun to send bets
}
