% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidraSearch.R
\name{sidraSearch}
\alias{sidraSearch}
\title{Search for Sidra Series}
\usage{
sidraSearch(description = NULL, code, view = TRUE, browse = FALSE)
}
\arguments{
\item{description}{A \code{character} argument. Function searches the Sidra metadata and prints results in a window.}

\item{code}{A numeric argument must be provided. The descriptions of the given table are returned.}

\item{view}{A \code{boolean}. The default is \code{TRUE}. If set to \code{FALSE}, the results are NOT going to be shown.}

\item{browse}{A \code{boolean}. If browse is set to \code{TRUE}, the description table opens in your browser for better visualization.}
}
\description{
Searches the Sidra databases for a series by its description or a given table descriptions.
}
\examples{
\dontrun{
sidraSearch(description = "pib")
sidraSearch(code = 1248)
}
}
\keyword{IBGE}
\keyword{sidra}
