% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_resid.R
\name{std_resid}
\alias{std_resid}
\title{Plot standardized residuals}
\usage{
std_resid(model, alpha = 0.05)
}
\arguments{
\item{model}{An \link[forecast]{Arima} or an \link[stats]{arima} object. The model.}

\item{alpha}{A \code{numeric} between 0 and 1. The significance level.}
}
\value{
Besides showing the plot, this function returns a \code{numeric vector} containing the standardized residuals.
}
\description{
Uses a model object to create a plot of standardized residuals. This model can be an \link[forecast]{Arima} or an \link[stats]{arima}. In a near future, this function will also accept objects returned by \link[BETS]{grnn.train}.
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
