% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_anuais.R
\docType{data}
\name{ts_anuais}
\alias{ts_anuais}
\title{Annual time series of BETS}
\format{A data frame with 34917 rows and 3 variables:
 \describe{
   \item{data}{Date of ts observation}
   \item{valor}{observation}
   \item{serie}{Code of time serie}
 }}
\source{
\url{http://portalibre.fgv.br/}
}
\description{
A dataset containing the annual time series
}
\keyword{data}

