% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_diarias.R
\docType{data}
\name{ts_diarias}
\alias{ts_diarias}
\title{Daily time series of BETS}
\format{A data frame with 2225839 rows and 3 variables:
 \describe{
   \item{data}{Date of ts observation}
   \item{valor}{observation}
   \item{serie}{Code of time serie}
 }}
\source{
\url{http://portalibre.fgv.br/}
}
\description{
A dataset containing the daily time series
}
\keyword{data}

