% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_semanais.R
\docType{data}
\name{ts_semanais}
\alias{ts_semanais}
\title{Weekly time series of BETS}
\format{A data frame with 4715 rows and 3 variables:
 \describe{
   \item{data}{Date of ts observation}
   \item{valor}{observation}
   \item{serie}{Code of time serie}
 }
  @source \url{http://portalibre.fgv.br/}
  @keywords data}
\description{
A dataset containing the weekly time series
}

