% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.notes.R
\name{add.notes}
\alias{add.notes}
\title{Add notes}
\usage{
add.notes(series.list, xlim, ylim, names = NULL, dec = 2)
}
\arguments{
\item{series.list}{A \code{ts object}}

\item{xlim}{A \code{vector}}

\item{ylim}{A \code{vector}}

\item{names}{A \code{character}}

\item{dec}{An \code{integer}}
}
\description{
Add notes
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

