% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.chart.R
\name{BETS.chart}
\alias{BETS.chart}
\title{Create a chart with BETS aesthetics}
\usage{
BETS.chart(ts, style = "normal", file = NULL, open = TRUE, lang = "en",
  params = NULL)
}
\arguments{
\item{ts}{A \code{character} or a \code{ts} object. A custom time series or the name of a pre-defined series. A complete list of names is under the 'Details' section.}

\item{style}{A \code{character}. Should the chart be made with Plotly (style = "plotly") or with R standard library (style = "normal")?}

\item{file}{A \code{character}. The whole path, including a custom name, for the output (an image file). The default value is NULL. If left to NULL, the chart will be rendered in the standard R plotting area.}

\item{open}{A \code{boolean}. Whether to open the file containing the chart.}

\item{lang}{A \code{character}. The language. For now, only 'en' (english) is available.}

\item{params}{A \code{list}. Parameters for drawing custom charts. See the 'details' section.}
}
\value{
If parameter \code{file} is not set by the user, the chart will be shown at the standard R ploting area. Otherwise, it is going to be saved on your computer.
}
\description{
Create a professional looking chart, using a pre-defined BETS series or a custom series.
}
\details{
\bold{Names of pre-defined charts:}

Business Cycle Dashboard ('plotly' style):

\tabular{lll}{
 VALUE \tab DESCRIPTION \tab CODE \cr
 \bold{'iie_br'} \tab Uncertainty Index \tab ST_100.0 \cr
 \bold{'sent_ind'} \tab Economic Sentiment Index (average between several confidence indexes) \tab (*) \cr
 \bold{'gdp_mon'} \tab GDP Monthly and Interanual Variation (last values) - GDP Monitor (FGV/IBRE) \tab (*) \cr
 \bold{'ei_vars'} \tab Economic Indicators (Leading and Coincident) monthly variation \tab (*) \cr
 \bold{'ei_comps'} \tab Economic Indicators (Leading and Coincident) components variation  \tab (*) \cr
 \bold{'lei'} \tab Leading Economic Indicator (LEI - FGV/IBRE with The Conference Board) \tab (*) \cr
 \bold{'cei'} \tab Coincident Economic Indicator (CEI - FGV/IBRE with the Conference Board)  \tab (*) \cr
 \bold{'gdp_vars'} \tab GDP components variation (whole series) - GDP Monitor (FGV/IBRE) \tab (*) \cr
 \bold{'misery_index} \tab Misery Index \tab 13522 plus 24369 \cr
 \bold{'gdp_comps'} \tab GDP components variation (last values) - GDP Monitor (FGV/IBRE) \tab (*) \cr
 \bold{'gdp_unemp'} \tab GDP monthly levels versus Unemployement Rate \tab 22109 and 24369 \cr
 \bold{'conf_lvl'} \tab Enterprises Confidence Index versus Consumers Confidence Index \tab (*) \cr
 \bold{'inst_cap'} \tab Installed Capacity Index \tab (*) \cr
 \bold{'lab_lead'} \tab Labor Leading Indicator \tab (*) \cr
 \bold{'lab_coin'} \tab Labor Coincident Indicator \tab (*) \cr
 \bold{'transf_ind'} \tab Transformation Industry Confidence Index (Expectations versus Present Situation) \tab (*) \cr
 \bold{'servc'} \tab Services Confidence Index (Expectations versus Present Situation) \tab (*) \cr
 \bold{'constr'} \tab Construction Confidence Index (Expectations versus Present Situation) \tab (*) \cr
 \bold{'retail'} \tab Retail Sellers Confidence Index (Expectations versus Present Situation) \tab (*) \cr
 \bold{'consm'} \tab Consumer Confidence Index (Expectations versus Present Situation) \tab (*) 
}

Macro Situation Dashboard ('normal' style):

\tabular{lll}{
 VALUE \tab DESCRIPTION \tab CODE \cr
 \bold{'ipca_with_core'} \tab National consumer price index (IPCA) - in 12 months and  Broad national consumer price index - Core IPCA trimmed means smoothed \tab 13522 and 4466 \cr
 \bold{'ulc'} \tab Unit labor cost - ULC-US$ - June/1994=100 \tab 11777 \cr
 \bold{'eap'} \tab Economically active population \tab 10810 \cr
 \bold{'cdb'} \tab Time deposits (CDB/RDB-preset) - Daily return (percentage) \tab 14 \cr
 \bold{'indprod'} \tab Prodcution Indicators (2012=100) - General	\tab 21859 \cr
 \bold{'selic'} \tab Interest rate - Selic accumulated in the month in annual terms (basis 252) \tab 4189 \cr
 \bold{'unemp'} \tab Unemployment rate - by metropolitan region (PNAD-C) \tab 10777\cr
 \bold{'vargdp'} \tab GDP - real percentage change in the year \tab 7326 
}

(*) Not available on BETS databases yet. But you can find it in .csv files saved under your BETS installation directory.

Custom Dashboards:

\enumerate{
\item{\code{type} A \code{character}. Either 'bar' or 'lines'. Whether to plot bars or lines. Works for main series, only.}
\item{\code{trend} A \code{boolean}. Default is \code{FALSE}. Set it to \code{TRUE} if the trend of the main series (parameter \code{ts}) is to be drawn.}
\item{\code{title}} A \code{character}. The plot's title.
\item{\code{subtitle}} A \code{character}. The plot's subtitle.
\item{\code{xlim}} A \code{numeric} vector. X axis limits 
\item{\code{ylim}} A \code{numeric} vector. Y axis limits
\item{\code{arrow}} A \code{character}. The orientation of the arrow pointing to the last value of the main series. Valid values are 'h' (horizontal) and 'v' (vertical). Doesn't work for 'plotly' type plots.
\item{\code{extra}} A \code{ts} object. A second series to be plotted.
\item{\code{extra.y2}} A \code{boolean}. Default is \code{FALSE}. Does the extra series require a second y axis? 
\item{\code{extra.arrow}} A \code{character}. The orientation of the arrow pointing to the last value of the extra series. Valid values are 'h' (horizontal) and 'v' (vertical). Doesn't work for 'plotly' type plots.
\item{\code{colors}} A \code{character} or \code{integer} vector. A vector of colors, one for each series. Trends will always be drawn in gray, its color can't be set. 
\item{\code{legend}} A \code{character} vector. The names of the series. Default is \code{NULL} (no legends).
\item{\code{legend.pos}} A \code{character}. The legend position. Possibile values are 'topleft','bottomleft', 'topright' and 'bottomright'.
\item{\code{codace}} A \code{boolean}. Default is \code{FALSE}. Include shaded areas for recessions, as dated by CODACE(**)? This feature only works if \code{style} is set to 'plotly'.
}

(**) Business Cycle Dating Committee (FGV/IBRE)
}
\note{
It is not yet possible to make charts of custom series.
}
\examples{

# BETS.chart(ts = "sent_ind", file = "animal_spirits", open = T)
# BETS.chart(ts = "gdp_mon", file = "gdp_mon.png", open = F)
# BETS.chart(ts = "misery_index")
# BETS.chart(ts = "transf_ind", file = "transf_ind.png", open = F)

}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
