% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.chart.R
\name{BETS.chart}
\alias{BETS.chart}
\title{Create a chart with a pre-defined BETS series}
\usage{
BETS.chart(alias, lang = "en", out = "png", file = NULL, start = c(2006,
  1), ylim = NULL, open = TRUE)
}
\arguments{
\item{alias}{A \code{character}. The alias of the chart. A complete list of available charts is under the 'Details' section.}

\item{lang}{A \code{character}. The language. For now, only 'en' (english) is available.}

\item{out}{A \code{character}. The format of the output, that is, the image file. Can be either 'pdf' or 'png'. 'pdf' is a better choice if you need high resolution images.}

\item{file}{A \code{character}. The whole path, including a custom name, for the output (an image file). The default value is 'graphs//parameter_alias' (the 'graphs' directory is under the BETS installation directory).}
}
\value{
If the parameter \code{file} is not set by the user, the chart will be placed in the 'graphs' directory, under the BETS installation directory.
}
\description{

}
\details{
\tabular{lll}{
 \code{alias} \tab Description \tab BETS code \cr
 'ipca_with_core' \tab National consumer price index (IPCA) - in 12 months and  Broad national consumer price index - Core IPCA trimmed means smoothed \tab 13522 and 4466
}
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

