% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.dashboard.R
\name{BETS.dashboard}
\alias{BETS.dashboard}
\title{Create a BETS custom dashboard}
\usage{
BETS.dashboard(type = "business_cycle", charts = "all", saveas = NA,
  parameters = NULL)
}
\arguments{
\item{type}{A \code{character}. The theme of the dashboard. The only three options, for the time being, is 'business_cycle', 'macro_situation' and 'custom'. Custom dashboards can be rendered with any given set of charts.}

\item{charts}{A \code{character} and/or \code{ts} object list. The charts to be added to a custom dashboard. Up to 16 charts are allowed, including pre-defined charts, identified by their codes (see \code{\link{BETS.chart}}). This will only work if parameter 'type' is set to 'custom'.}

\item{saveas}{A \code{character}. A path and a name for the dashboard file (a .pdf file). If this parameter is not provided, the dashboard will be saved inside the 'dashboards' folder, under the BETS installation directory.}

\item{parameters}{A \code{list}. A list of parameters. See the 'Details' section for a description of these parameters for each type of dashboard.}
}
\value{
A .pdf file (the dashboard)
}
\description{
Generate thematic dashboards using a selection of BETS time series and charts. For now, themes and charts are pre-defined.
}
\details{
\bold{Macro Situation and Custom Dashboard Parameters} 

\tabular{ll}{
\code{text}\tab The text to be printed in the dashboard. Separate paragraphs with two backslashes 'n' and pages with '##'. There are no other syntax rules.\cr
\code{author}\tab The author's name.\cr
\code{email}\tab The author's email.\cr
\code{url}\tab The author's webpage.\cr
\code{logo}\tab The author's business logo.\cr
}


\bold{Additional Custom Dashboard Parameters}

\tabular{ll}{
\code{style} \tab A \code{character}. The style of the charts. As in \code{BETS.chart}, can be either \code{'plotly'} or \code{'normal'}.\cr
\code{charts.opts} \tab A \code{list} of parameters lists, one for each chart. Parameters are specified in \code{\link{BETS.chart}} \cr
}
}
\examples{

# BETS.dashboard()
# BETS.dashboard(saveas = "survey.pdf")
# BETS.dashboard(type = "macro_situation")

}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
