% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.dashboard.R
\name{BETS.dashboard}
\alias{BETS.dashboard}
\title{xxx}
\usage{
BETS.dashboard(type = "inflation", parameters = NULL, saveas = NA)
}
\arguments{
\item{model}{xxx}

\item{params}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

