% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.dashboard.R
\name{BETS.dashboard}
\alias{BETS.dashboard}
\title{Create a BETS custom dashboard}
\usage{
BETS.dashboard(type = "inflation", country = "BR", parameters = NULL,
  saveas = NA)
}
\arguments{
\item{type}{A \code{character}. The theme of the dashboard. The only option, for the time being, is 'inflation'.}

\item{country}{A \code{character}. The country from which data must drawn. For now, BETS only accepts 'BR' (Brazil)}

\item{parameters}{A \code{list}. A list of parameters. See the 'Details' section for a description of these parameters for each type of dashboard.}

\item{saveas}{A \code{character}. A path and a name for the dashboard file (a .pdf file). If this parameter is not provided, the dashboard will be saved inside the 'dashboards' folder, under the BETS installation directory.}
}
\value{
A .pdf file (the dashboard)
}
\description{
Generate thematic dashboards using a selection of BETS time series and charts. For now, themes and charts are pre-defined.
}
\details{
\bold{Inflation Dashboard Parameters} 

\itemize{
\item{ \code{text} - The text to be printed in the dashboard. Separate paragraphs with two backslashes 'n' and pages with '##'. There are no other syntax rules.}
\item{ \code{author} - The author's name.}
\item{ \code{email} - The author's email.}
\item{ \code{url} - The author's webpage.}
\item{ \code{logo} - The author's business logo.}
}
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}

