% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.dashboard.R
\name{BETS.dashboard}
\alias{BETS.dashboard}
\title{Create a BETS custom dashboard}
\usage{
BETS.dashboard(type = "sentiment", saveas = NA)
}
\arguments{
\item{type}{A \code{character}. The theme of the dashboard. The only option, for the time being, is 'sentiment'.}

\item{saveas}{A \code{character}. A path and a name for the dashboard file (a .pdf file). If this parameter is not provided, the dashboard will be saved inside the 'dashboards' folder, under the BETS installation directory.}
}
\value{
A .pdf file (the dashboard)
}
\description{
Generate thematic dashboards using a selection of BETS time series and charts. For now, themes and charts are pre-defined.
}
\examples{

# BETS.dashboard()
# BETS.dashboard(saveas = "survey.pdf")

}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
