% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.deflate.R
\name{BETS.deflate}
\alias{BETS.deflate}
\title{Deflate a time series}
\usage{
BETS.deflate(ts, deflator, type = "index")
}
\arguments{
\item{ts}{A \code{ts} object. The time series to be deflated.}

\item{deflator}{A \code{ts} object. The deflator series.}

\item{type}{A \code{character}. Can be either \code{'index'}, \code{'point.perc'} (for point percentage) or \code{'perc'} (for percentage).}
}
\value{
The deflated series.
}
\description{
Deflate a time series using a deflator series. The deflator can be an index, a percentage or a point percentage series.
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
