% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.sidra.search.R
\name{BETS.sidra.search}
\alias{BETS.sidra.search}
\title{Search for Sidra Series}
\usage{
BETS.sidra.search(x, browse = FALSE)
}
\arguments{
\item{x}{Either a character or a numeric. If character, function searches the Sidra metadata. If a numeric argument is provided the descriptions of the given table are seached .}

\item{browse}{Logical. If browse is set to TRUE, the description table opens in your browser for better visualization.}
}
\description{
Searches the Sidra databases for a series by its description or a given table descriptions.
}
\examples{
\dontrun{
BETS.sidra.search("pib")
BETS.sidra.search(1248)
}
}
\keyword{IBGE}
\keyword{sidra}
