% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSget.R
\name{BETSget}
\alias{BETSget}
\title{Extrair series temporais do Banco central do Brasil, IBGE  e do 
Instituto brasileiro de economia da Fundacao Getulio Vargas.}
\usage{
BETSget(codigo)
}
\arguments{
\item{code}{que referencia a serie temporal}
}
\value{
Um objeto do tipo TS
}
\description{

}
\note{
Insira o valor ao argumento code como character. 
Devido ao grande volume de dados por parte das obeservacoes 
da series temporais, pode levar agum tempo para o algoritmo resgatar
os valores, no entanto esse tempo e breve nao passando de 90 segundos.
}
\examples{

 #Aconselha-se a usar a funcao GETsearch() para 
 #localizar a serie desejada, obtendo assim o 
 #codigo da serie.
 
 
 #serie anual: GDP at constant last year prices in R$
 BETSget(1208)
 #International reserves - Cash concept 
 x<-BETSget("3543")
 plot(x)
 #Exchange rate - Free - United States dollar(purchase)
 k=BETSget(3691)
 library(seasonal)
 m <- seas(k)
 plot(m)
 
}
\references{

}

