% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsave.spss.R
\name{BETSsave.spss}
\alias{BETSsave.spss}
\title{Export a time series to SPSS}
\usage{
BETSsave.spss(code = NULL, data = NULL, file.name = "series")
}
\arguments{
\item{code}{An \code{integer}. The unique identifier of the series within the BETS database.}

\item{data}{A \code{data.frame} or a \code{ts}. Contains the data to be written. If \code{data} is supplied, the BETS database will not be searched.}

\item{file.name}{A \code{character}. The name of the output file. The default is 'series.spss'.}
}
\value{
None
}
\description{
Writes a time series to a .spss (SPSS) file.
}

