% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETSsearch.R
\name{BETSsearch}
\alias{BETSsearch}
\title{Search for a brazilian economic time series}
\usage{
BETSsearch(name,Source,periodicity,unit,code,view=TRUE,data)
}
\arguments{
\item{name}{Descricao ou parte da descricao da serie temporal}

\item{Source}{Fonte de onde e produzida}

\item{periodicity}{Frequencia da serie temporal}

\item{unit}{Unidade da serie temporal}

\item{code}{Codigo que referencia a serie temporal}

\item{view}{TRUE para visualizar os resultado diretamente}

\item{data}{Em qual base sera feita a busca}
}
\value{
Objeto do tipo list de tamanho 5  com  estetica de um data.frame.

\itemize{

\item{Codes }{Codigos das series.}
\item{Descripition}{Descricao das seires.}
\item{Periodicity}{Periodicidade das seires.}
\item{source}{Fonte des series.}
\item{unit}{Unidade da serie.}
}


}
\description{

}
\note{

}
\examples{

# The function is currently defined as
BETSsearch(name="sales")
#BETS-package: 55 of 12981 time series !
searchBETS(code= 4500)
#BETS-package: DONE!
BETSsearch(Source="Denor")
#BETS-package: 1 of 12981 time series !
BETSsearch(periodicity="A")
#BETS-package: 2308 of 12981 time series!

#Como mencionado em 'Note', segue os 
#exemplos de pesquisas combinadas:
BETSget(name="production",periodicity="M",view=FALSE)
BETSget(Source=BCB,periodicity="M",view=TRUE)
BETSget(name="production",Source="IBGE",view=TRUE)
BETSget(unit="index",periodicity="M",view=FALSE)

}
\references{
Banco Central do Brasil:
Time Series Management System - v2.1 (Public module)
}
\keyword{search}

