% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.series.R
\name{check.series}
\alias{check.series}
\title{Check series}
\usage{
check.series(ts, message = NULL)
}
\arguments{
\item{ts}{A \code{ts object}}

\item{message}{A \code{character}}
}
\description{
Check series in BETS dataset
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
