% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.cdb.R
\name{draw.cdb}
\alias{draw.cdb}
\title{Create a chart of the Time Deposits time series}
\usage{
draw.cdb()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 14
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
