% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.cdb.R
\name{draw.cdb}
\alias{draw.cdb}
\title{Create a chart of the Time Deposits time series}
\usage{
draw.cdb(start = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{start}{A \code{character}. The stating period of the series.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{xlim}{A \code{numeric vector}. x axis limits.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 14
}

