% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.eap.R
\name{draw.eap}
\alias{draw.eap}
\title{Create a chart of the Economically Active Population time series}
\usage{
draw.eap()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 10810
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
