% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.indprod.R
\name{draw.indprod}
\alias{draw.indprod}
\title{Create a chart of the Production Indicators time series}
\usage{
draw.indprod()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 21859
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
