% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ipca.R
\name{draw.ipca}
\alias{draw.ipca}
\title{xxx}
\usage{
draw.ipca(file, start = NULL, ylim = NULL, open = TRUE)
}
\arguments{
\item{ts}{xxx}

\item{lag.max}{xxx}

\item{mode}{xxx}

\item{ci}{xxx}
}
\value{
xxx
}
\description{
xxxx
}

