% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ipca.R
\name{draw.ipca}
\alias{draw.ipca}
\title{Create a chart of the National Consumer Price Index time series}
\usage{
draw.ipca(start = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{start}{A \code{character}. The stating period of the series.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{xlim}{A \code{numeric vector}. x axis limits.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 13522 (NCPI), along with series 4466 (NCPI core)
}

