% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.selic.R
\name{draw.selic}
\alias{draw.selic}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.selic()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}
}
