% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.selic.R
\name{draw.selic}
\alias{draw.selic}
\title{Create a chart of the Base Interest Rate (SELIC) time series}
\usage{
draw.selic(start = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{start}{A \code{character}. The stating period of the series.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{xlim}{A \code{numeric vector}. x axis limits.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 4189
}

