% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ulc.R
\name{draw.ulc}
\alias{draw.ulc}
\title{Create a chart of the Unitary Labor Cost time series}
\usage{
draw.ulc()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 11777
}
