% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.ulc.R
\name{draw.ulc}
\alias{draw.ulc}
\title{Create a chart of the Unitary Labor Cost time series}
\usage{
draw.ulc(start = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{start}{A \code{character}. The stating period of the series.}

\item{ylim}{A \code{numeric vector}. Y axis limits.}

\item{xlim}{A \code{numeric vector}. x axis limits.}
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 11777
}

