% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.unemp.R
\name{draw.unemp}
\alias{draw.unemp}
\title{Create a chart of the Open Unemployment Rate time series}
\usage{
draw.unemp()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 10777
}
