% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.vargdp.R
\name{draw.vargdp}
\alias{draw.vargdp}
\title{Create a chart of the Real Percentage Change of GDP in the Year time series}
\usage{
draw.vargdp()
}
\value{
An image file is saved in the 'graphs' folder, under the BETS installation directory.
}
\description{
Creates a plot of series 7326
}
