% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fbref_data.R
\name{get_fbref_data}
\alias{get_fbref_data}
\title{get fbref data}
\usage{
get_fbref_data(seasons)
}
\arguments{
\item{seasons}{a vector of yearst}
}
\value{
a tibble containing data.
}
\description{
get fbref data
}
