% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIE-Br-Mercado.R
\name{IIE-Br-Market}
\alias{IIE-Br-Market}
\title{IIE-Br-Market}
\format{A \code{\link[stats]{ts}} object with 185 observations.}
\source{
\href{http://portalibre.fgv.br/}{Brazilian Institute of Economics (FGV/IBRE)} 
and \href{http://www.bcb.gov.br/?english}{Central Bank of Brasil (BCB)}
}
\description{
O \code{IIE-Br-Market} visa mensurar variabilidade do mercado acionaario
IIE-Br-Market.

Brasileiro e seu grau de risco, ou seja, diferente IIE-Br-expectativa busca
mensurar a variabilidade do sentimento do mercado no tempopresente.

O \code{IIE-Br-Market}  e composto pela volatilidade dos ultimos 21 dias 
decalculo dos precos das ações do IBOVESPA e opremio de cinco anos do Credit 
Swap Default.
}
\author{
Pedro G. C. Ferreira  \email{pedro.guilherme@fgv.br},
        Anna Carolina S. Barros \email{anna.barros@fgv.br},
        Bruno R de Miranda Neto \email{bruno.neto@fgv.br},
        Itaiguara de Oliveira Bezerra  \email{itaiguara.bezerra@fgv.br}.
}
\references{
A shiny app with IIE-Br-expectations plots can be found 
             \href{here}{https://pedroferreira.shinyapps.io/incerteza/}
}
\keyword{IIE-Br-Market}

