% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Format and show a console message.}
\usage{
msg(..., skip_before = TRUE, skip_after = FALSE, warn = FALSE)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[base]{message}}}

\item{skip_before}{A \code{boolean}. Indicates if a line should be skipped before the message.}

\item{skip_after}{A \code{boolean}. Indicates if a line should be skipped after the message.}

\item{warn}{A \code{boolean}. Indicates whether a warning should be thrown.}
}
\value{
None
}
\description{
Customizes a message and shows it in the console.
}
\author{
Talitha Speranza \email{talitha.speranza@fgv.br}, Jonatha Azevedo \email{jonatha.azevedo@fgv.br}
}
