% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_team_names.R
\name{replace_team_names}
\alias{replace_team_names}
\title{replaces team names in data frame}
\usage{
replace_team_names(data, team1, team2, pattern, replacement)
}
\arguments{
\item{data}{account name}

\item{team1}{account name}

\item{team2}{account name}

\item{pattern}{account name}

\item{replacement}{account name}
}
\description{
replaces team names in data frame
}
\examples{
hist_buch_data \%>\% replace_team_names(home, away, team_dictionary()$buch_name, team_dictionary()$fbref_name)
}
