% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Significancia dos parametro de um modelo Arima}
\usage{
t_test(arima_model, n_x = 0)
}
\arguments{
\item{arima_model}{Modelo arima utilizado}

\item{n_x}{Numero de variaveis exogenas}
}
\value{
Objeto do tipo \code{list}
}
\description{
Realiza o teste de significancia dos parametro de um modelo 
ARIMA.
}
\details{

}
\author{
Daiane Mattos \email{daiane.mattos@fgv.br}
}

